requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
MouseInfo==0.1.3
mss==9.0.1
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-Quartz==9.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytesseract==0.3.10
pytweening==1.0.7
requests==2.31.0
rubicon-objc==0.4.6
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.6
wsproto==1.2.0
